+++ mono-traversable/src/Data/Containers.hs
@@ -616,6 +616,8 @@ instance Ord key => IsMap (Map.Map key value) where
     {-# INLINE mapWithKey #-}
     omapKeysWith = Map.mapKeysWith
     {-# INLINE omapKeysWith #-}
+    filterMap = Map.filter
+    {-# INLINE filterMap #-}
 
 #if MIN_VERSION_containers(0, 5, 0)
 -- | This instance uses the functions from "Data.HashMap.Strict".
@@ -653,6 +655,8 @@ instance (Eq key, Hashable key) => IsMap (HashMap.HashMap key value) where
     --unionsWith = HashMap.unionsWith
     --mapWithKey = HashMap.mapWithKey
     --mapKeysWith = HashMap.mapKeysWith
+    filterMap = HashMap.filter
+    {-# INLINE filterMap #-}
 
 #if MIN_VERSION_containers(0, 5, 0)
 -- | This instance uses the functions from "Data.IntMap.Strict".
@@ -703,6 +707,8 @@ instance IsMap (IntMap.IntMap value) where
     omapKeysWith = IntMap.mapKeysWith
     {-# INLINE omapKeysWith #-}
 #endif
+    filterMap = IntMap.filter
+    {-# INLINE filterMap #-}
 
 instance Eq key => IsMap [(key, value)] where
     type MapValue [(key, value)] = value
@@ -736,6 +742,12 @@ class (SetContainer set, Element set ~ ContainerKey set) => IsSet set where
     -- | Convert a set to a list.
     setToList :: set -> [Element set]
 
+    -- | Filter values in a set.
+    --
+    -- @since 1.0.12.0
+    filterSet :: (Element set -> Bool) -> set -> set
+    filterSet p = setFromList . filter p . setToList
+
 instance Ord element => IsSet (Set.Set element) where
     insertSet = Set.insert
     {-# INLINE insertSet #-}
@@ -747,6 +759,8 @@ instance Ord element => IsSet (Set.Set element) where
     {-# INLINE setFromList #-}
     setToList = Set.toList
     {-# INLINE setToList #-}
+    filterSet = Set.filter
+    {-# INLINE filterSet #-}
 
 instance (Eq element, Hashable element) => IsSet (HashSet.HashSet element) where
     insertSet = HashSet.insert
@@ -759,6 +773,8 @@ instance (Eq element, Hashable element) => IsSet (HashSet.HashSet element) where
     {-# INLINE setFromList #-}
     setToList = HashSet.toList
     {-# INLINE setToList #-}
+    filterSet = HashSet.filter
+    {-# INLINE filterSet #-}
 
 instance IsSet IntSet.IntSet where
     insertSet = IntSet.insert
@@ -771,6 +787,8 @@ instance IsSet IntSet.IntSet where
     {-# INLINE setFromList #-}
     setToList = IntSet.toList
     {-# INLINE setToList #-}
+    filterSet = IntSet.filter
+    {-# INLINE filterSet #-}
 
 
 -- | Zip operations on 'MonoFunctor's.